+++ src/Utils.cs
@@ -271,6 +271,7 @@ namespace DCFApixels.WhimTex
         private static Material makeSeamlessMaterial;
         private static Material noiseMaterial;
         private static Material gradientMaterial;
+        private static Material fillUvMaterial;
         private static Material shapeMaterial;
         private static Material effectCacheMaterial;
 
@@ -288,6 +289,7 @@ namespace DCFApixels.WhimTex
         public static Material MakeSeamless => GetOrCreate(ref makeSeamlessMaterial, "Hidden/TextureCompositor/MakeSeamless");
         public static Material Noise => GetOrCreate(ref noiseMaterial, "Hidden/TextureCompositor/Noise");
         public static Material Gradient => GetOrCreate(ref gradientMaterial, "Hidden/TextureCompositor/Gradient");
+        public static Material FillUv => GetOrCreate(ref fillUvMaterial, "Hidden/TextureCompositor/FillUv");
         public static Material Shape => GetOrCreate(ref shapeMaterial, "Hidden/TextureCompositor/Shape");
         public static Material EffectCache => GetOrCreate(ref effectCacheMaterial, "Hidden/TextureCompositor/EffectCache");
         public static Material Transform => GetOrCreate(ref transformMaterial, "Hidden/TextureCompositor/Transform");
@@ -318,6 +320,8 @@ namespace DCFApixels.WhimTex
 
         private static void Dispose()
         {
+            if (fillUvMaterial != null) UnityEngine.Object.DestroyImmediate(fillUvMaterial);
+            fillUvMaterial = null;
             if (gradientMaterial != null) UnityEngine.Object.DestroyImmediate(gradientMaterial);
             gradientMaterial = null;
             if (shapeMaterial != null) UnityEngine.Object.DestroyImmediate(shapeMaterial);